fix: 로그인 직후 폰트 미반영 문제 수정
